agent_template_57978/requirements.txt
python-dotenv
langchain-openai
langchain-core
langchain-community
langgraph